backend/pyproject.toml
[project]
name = "butterfly-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn",
    "litellm",
    "python-dotenv",
    "pydantic-ai[logfire]>=0.6.2",
    "nest-asyncio>=1.6.0",
    "logfire>=4.2.0",
    "runware>=0.4.18",
]

[dependency-groups]
dev = [
    "pytest-asyncio>=1.1.0",
]
